fix: e2e tests
