+++ apps/meteor/tests/e2e/page-objects/fragments/sidebar.ts
@@ -1,7 +1,5 @@
 import type { Locator, Page } from '@playwright/test';
 
-import { expect } from '../../utils/test';
-
 export class Sidebar {
 	private readonly page: Page;
 
@@ -14,20 +12,12 @@ export class Sidebar {
 		return this.page.getByRole('navigation', { name: 'sidebar' });
 	}
 
-	get sidebarSearchSection(): Locator {
-		return this.sidebar.getByRole('search');
-	}
-
-	get btnRecent(): Locator {
-		return this.sidebarSearchSection.getByRole('button', { name: 'Recent' });
-	}
-
 	get channelsList(): Locator {
 		return this.sidebar.getByRole('list', { name: 'Channels' });
 	}
 
-	get searchList(): Locator {
-		return this.sidebar.getByRole('search').getByRole('list', { name: 'Channels' });
+	getSearchRoomByName(name: string) {
+		return this.channelsList.getByRole('link', { name });
 	}
 
 	get firstCollapser(): Locator {
@@ -38,35 +28,8 @@ export class Sidebar {
 		return this.channelsList.getByRole('listitem').first();
 	}
 
-	get searchInput(): Locator {
-		return this.sidebarSearchSection.getByRole('searchbox');
-	}
-
 	async escSearch(): Promise<void> {
 		await this.page.keyboard.press('Escape');
-	}
-
-	async waitForChannel(): Promise<void> {
-		await this.page.locator('role=main').waitFor();
-		await this.page.locator('role=main >> role=heading[level=1]').waitFor();
-		const messageList = this.page.getByRole('main').getByRole('list', { name: 'Message list', exact: true });
-		await messageList.waitFor();
-
-		await expect(messageList).not.toHaveAttribute('aria-busy', 'true');
-	}
-
-	async typeSearch(name: string): Promise<void> {
-		return this.searchInput.fill(name);
-	}
-
-	getSearchRoomByName(name: string): Locator {
-		return this.searchList.getByRole('link', { name });
-	}
-
-	async openChat(name: string): Promise<void> {
-		await this.typeSearch(name);
-		await this.getSearchRoomByName(name).click();
-		await this.waitForChannel();
 	}
 
 	async markItemAsUnread(item: Locator): Promise<void> {
